Fix for add_special_tokens=False
Sometimes squeeze() could get [1.0] as input, transforming it into a 0-dim scalar tensor.

Using tolist() on a scalar tensor does not return a list, it returns a number, causing breakage. It's a little strange that tolist() doesn't always return a list, but that's how pytorch devs decided to have it.

Using flatten() instead of squeeze() prevents this. Flattening always returns a 1-dim tensor.